LPS-201076 Move BundleManagerUtil from service to api module and replace each use of BundleManager or different BundleManagerUtil.
