+++ modules/apps/marketplace/marketplace-api/src/main/resources/com/liferay/marketplace/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
